feat(git-ai): Update git commit message template
Update the git commit message template to include instructions on how to include the diff content in the body of the commit message. Also specify that only one <type> should be used and that the commit message should not explain why a commit was made.

Details:
- Update git commit message template
- Include instructions on how to include the diff content in the body of the commit message
- Specify that only one <type> should be used
- Specify that the commit message should not explain why a commit was made

Commit written by git-ai. :)