fix(build): Update build script to add changes and run the generated script (#10)
Updating the build script in the package.json file to include the command `git add.`
after running the `bun build` command. This will add all the changes to the Git
repository before running the `bun run` command to execute the generated script.

This change ensures that all the changes in the build directory are properly added
to the Git repository before running the generated script.